Make scalar type hints nullable.
Since the interfaces defined in the psr library do not have type hints,
the scalar type hints for method parameters must be nullable.  If they
are not, a runtime error occurs stating any implementation is not
consistent with the interface, even when the implementation is
consistent with the hhi interface.